requirements-dev.txt
-r requirements.txt
black==21.8b
coverage>=6.3.1,<7
PyYAML==5.4.1
responses>=0.18.0
pre-commit>=2.17.0,<3